requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
pandas>=2.1.0
boto3>=1.34.0
pgvector>=0.2.0
requests>=2.31.0
tqdm>=4.66.0
